optee: include ftpm ta by default
fTPM can work as a proof of concept using REE_FS, i.e. by storing an
encrypted blob on the linux rootfs, which is protected by an
soc-internal key only accessible by optee-os itself.

The TA is built as a 32-bit application to minimize code size and fit
within the dedicated boot sectors.